Revert "refactor(test): move has_display fixture to conftest.py for reuse across test modules"
This reverts commit 5a861e748f45b35c9d044f5783a987b36a46ee2b.
